------------------------------------------------------------------------ r3696 | stial01 | 2010-10-12 11:30:09 -0400 (Tue, 12 Oct 2010) | 6 lines
SIR 121123 Long Ids

Description of changes:
Windows changes for new functions in cuid.c

------------------------------------------------------------------------
r3697 | loera01 | 2010-10-12 13:00:12 -0400 (Tue, 12 Oct 2010) | 10 lines

Changes made for bug 124578 in main didn't include tightening up the
treatment of of the outLen and outStr variables in SQLBrowseConnect() and
SQLBrowseConnectW().  I'm adding the recommended changes.

I also forgot to include a change I made to the driver in SQLDisconnect().
The bcConnectCalled boolean field in the connection handle needs to be set to
FALSE in so that SQLBrowseConnect_InternalCall() knows when the first
invocation of SQLBrowseConnect() is made.

------------------------------------------------------------------------