------------------------------------------------------------------------ r3697 | loera01 | 2010-10-12 13:00:12 -0400 (Tue, 12 Oct 2010) | 10 lines
Changes made for bug 124578 in main didn't include tightening up the
treatment of of the outLen and outStr variables in SQLBrowseConnect() and
SQLBrowseConnectW().  I'm adding the recommended changes.

I also forgot to include a change I made to the driver in SQLDisconnect().
The bcConnectCalled boolean field in the connection handle needs to be set to
FALSE in so that SQLBrowseConnect_InternalCall() knows when the first
invocation of SQLBrowseConnect() is made.

------------------------------------------------------------------------
r3698 | miket | 2010-10-12 13:08:56 -0400 (Tue, 12 Oct 2010) | 8 lines

Trac Number: #584

Bug Numbers: 122403 124542

Description of changes:
Add column encryption support for genxml/xmlimport. Also fix xfwrite
for MODIFY ENCRYPT PASSPHRASE to avoid extraneous \p\g.

------------------------------------------------------------------------