------------------------------------------------------------------------ r3702 | frima01 | 2010-10-14 02:15:18 -0400 (Thu, 14 Oct 2010) | 4 lines
Add enabling 8k page cache to iisudbms upgrade section
because it is needed for demodb tables.

------------------------------------------------------------------------
r3703 | gupsh01 | 2010-10-14 11:00:19 -0400 (Thu, 14 Oct 2010) | 18 lines

Rename operation on a table or a column succeeds
if there are dependent views on the table which
are not owned by the table owner. It is expected
that the error message E_US249D and E_US249E were
given out, for table and column rename respectively.

The problem is that the non owner views/procedures
are listed in iipriv catalog and are not available
from iidbdepends catalog. Scan iipriv and make
sure that these dependencies are reported.

Also fixed error messages text for
a) ALTER TABLE usage message.
b) Report owner name in views and procedures
dependent on table/column being renamed.
c) Report correct column name for column rename.

------------------------------------------------------------------------