------------------------------------------------------------------------ r3703 | gupsh01 | 2010-10-14 11:00:19 -0400 (Thu, 14 Oct 2010) | 18 lines
Rename operation on a table or a column succeeds
if there are dependent views on the table which
are not owned by the table owner. It is expected
that the error message E_US249D and E_US249E were
given out, for table and column rename respectively.

The problem is that the non owner views/procedures
are listed in iipriv catalog and are not available
from iidbdepends catalog. Scan iipriv and make
sure that these dependencies are reported.

Also fixed error messages text for
a) ALTER TABLE usage message.
b) Report owner name in views and procedures
dependent on table/column being renamed.
c) Report correct column name for column rename.

------------------------------------------------------------------------
r3704 | loera01 | 2010-10-14 20:00:12 -0400 (Thu, 14 Oct 2010) | 7 lines

Cross integration of change 507928 from main.

Fix memory leak in QRY_SETD case in odbc_query_sm(), the call to
PrepareSqldaAndBuffer() should only allocate sqlda if the query type
is QST_PREPARE--not the fetch buffer, too.

------------------------------------------------------------------------