------------------------------------------------------------------------ r3705 | smeke01 | 2010-10-15 04:45:20 -0400 (Fri, 15 Oct 2010) | 45 lines
Review Board IP  http://reviewboard.ingres.prv/r/6276

Cross integration of change 506145

In opc_hash_materialize() set up call to new function adu_norm_date_hash() to
normalise ingresdate (DB_DTE_TYPE) key columns so that hashing works. Also use
the same method as for varchars to make sure that the original un-normalised
date is passed up the query tree.

New function adu_norm_date_hash() takes an ingresdate and makes the necessary
adjustments to the data to make the ingresdate work correctly with hash. Most
of the work is done by calling existing functions. There are 3 kinds of date
that do not need to be normalised by this function: these are copied to the
output parameter unchanged. This is to keep the performance impact to a
minimum. The kinds of date that are put on this fast path are:

1. Absolute dates with time-spec.
2. Intervals specified only in days, or of less than a day.
3. The NULL date value.

Intervals are handled by retrieving the 'month days' equivalent of the interval,
and storing that value in the days fields (.dn_highday and .dn_lowday). The
dn_year and dn_month fields are zeroed, and the dn_status updated accordingly.
Note that the sequence of transformations deliberately follows that of
ade_4date_cmp(), which is the function used to compare dates eg by FSM join.

The handling of the operator in opc_hash_materialize() could be made slicker
by creating a single operator for hash normalisation, and creating instances
of it differentiated by data type.

Added Functions:

   None.

Removed Functions:

   None.

CL Interface changes:

   None.

------------------------------------------------------------------------
r3706 | hanal04 | 2010-10-15 06:00:45 -0400 (Fri, 15 Oct 2010) | 9 lines

Cross integration of change 507938

Review Board IP  http://reviewboard.ingres.prv/r/6304

Include systypes.h before wchar.h in iicommon.h to avoid build
failures on HP Itanium.

------------------------------------------------------------------------