------------------------------------------------------------------------ r3723 | maspa05 | 2010-10-19 05:00:22 -0400 (Tue, 19 Oct 2010) | 49 lines
Review Board IP  http://reviewboard.ingres.prv/r/6310

Stack was being trashed by the output of a BYTE/VARBYTE value for a
PARM/PARMEXEC under SC930.

This is because the buffer we use for the output string is size for the
maximum size of the datatype but we're writing it as "XX " for each hex digit
i.e. we'd need our current buffer x 3.

Rather than simply increase this buffer again, causing extra stack space to
be used, we now use adu_sc930prtdataval to write out the PARM/PARMEXEC lines.
Renamed adu_valuetomystr to adu_sc930prtdataval. It was/is only ever used by
SC930 to format DB_DATA_VALUE values for output to the SC930 trace. The
change of name reflects the fact that as well as format the value it now
outputs the value to the SC930 trace file.

So what used to happen would be the caller would allocate a char buffer,
call adu_valuetomystr with the dbdv value, add the resulting string into
another buffer with the other parts of the line ("PARM:... etc") and then
call ult_print_tracefile.

This is replaced with a single call to adu_sc930prtdataval. Which means we
now pass the file pointer, parameter name and number and the SC930 message
type. adu_sc930prtdataval first calles ult_print_tracefile to output the
first part of the line and then formats the dbdv as before except that now
it writes to the file instead of returning a string.

Changed ult_print_tracefile so that PARM/PARMEXEC do not have a newline at the
end.

Defined SC930_DATAVAL_BUFSIZE for the rough size of segment that we output the
strings in as well as the buffer for any string operations in
adu_sc930prtdataval.

The upshot of this is that we no longer need to allocate char buffers of
DB_MAXSTRING+ in the calling functions. Which means we should be using much
less stack and holding on to it for less time. Even within adu_sc930prtdataval
we only allocate the very large buffers if the data-type needs it, not for
every case.

Amended print_sc930_info in scsqncr.c, pst_cpdata and print_qry_buffer in
psqparse.c to call adu_adu_sc930prtdataval instead of ult_print_tracefile.

Whilst changing adu_sc930prtdataval added the output for DB_LOGKEY_TYPE
(object key) which was missing.

This change does not affect the SC930 format version which remains 8.

------------------------------------------------------------------------