------------------------------------------------------------------------ r3742 | maspa05 | 2010-10-25 10:45:30 -0400 (Mon, 25 Oct 2010) | 35 lines
Review Board IP  http://reviewboard.ingres.prv/r/6469

Cross integration of change 508100

iimonitor output of Query/Last Query was corrupted where the query text was
the maximum possible (ER_MAX_LEN).

This was because the string sent from the DBMS for output is "   Last Query:"
+query text so it could be upto ER_MAX_LEN+16. This is formatted with
TRformat where the first parameter (output function) is NULL. If the size of
the output exceeds the size of the buffer TRformat fills the buffer then calls
the output function, then starts filling the buffer again and so on - so in
this case you end up with the last few characters overwriting the first few,
hence the 'corrupted' output.

Changed iimonitor to increase the size of the buffer to ER_MAX_LEN+20. Also
we already reset length of the string to limit it to the size of the buffer,
moved this so that it happens before the call to TRformat.

Added Functions:

   None.

Removed Functions:

   None.

CL Interface changes:

   None.

------------------------------------------------------------------------