------------------------------------------------------------------------ r3742 | maspa05 | 2010-10-25 10:45:30 -0400 (Mon, 25 Oct 2010) | 35 lines
Review Board IP  http://reviewboard.ingres.prv/r/6469

Cross integration of change 508100

iimonitor output of Query/Last Query was corrupted where the query text was
the maximum possible (ER_MAX_LEN).

This was because the string sent from the DBMS for output is "   Last Query:"
+query text so it could be upto ER_MAX_LEN+16. This is formatted with
TRformat where the first parameter (output function) is NULL. If the size of
the output exceeds the size of the buffer TRformat fills the buffer then calls
the output function, then starts filling the buffer again and so on - so in
this case you end up with the last few characters overwriting the first few,
hence the 'corrupted' output.

Changed iimonitor to increase the size of the buffer to ER_MAX_LEN+20. Also
we already reset length of the string to limit it to the size of the buffer,
moved this so that it happens before the call to TRformat.

Added Functions:

   None.

Removed Functions:

   None.

CL Interface changes:

   None.

------------------------------------------------------------------------
r3743 | schka01 | 2010-10-25 12:36:51 -0400 (Mon, 25 Oct 2010) | 27 lines

Fix a few Mac OSX build problems encountered while trying to
debug on OSX 10.6.2 (latest xcode, gcc 4.2.1):

- precheck_functions was improperly checking for Xerces,
  producing warning messages.  (not just int_osx, happening on
  linux as well if Xerces not installed as a system package.)

- OSX linker apparently requires ranlib info on .a archives,
  or it refuses to link.

- ls -lo seems to be broken, causing iifilsiz to not work,
  causing config setup (iigenres) to fail.  Generalize
  iifilsiz to work around the problem.

- many compile/link errors in sig/dp/dp.  This isn't part
  of the distribution, but there's no harm in having it
  compile and link properly.

- add a -test flag to mksaveset so that createdbms can
  operate in the face of missing files.  The missing files
  were in sig caused by user error, but HOQA doesn't depend
  on sig, and it's painful to not be able to createdbms and
  run HOQA because of some unrelated problem.
  Note that I am not at the moment submitting a createdbms
  change;  createdbms can be changed locally if needed.

------------------------------------------------------------------------