------------------------------------------------------------------------ r3745 | maspa05 | 2010-10-26 12:45:22 -0400 (Tue, 26 Oct 2010) | 18 lines
Review Board IP  http://reviewboard.ingres.prv/r/6482

Original fix for bug 124551 may not output the whole value for byte, varbyte,
nchar or nvarchar data values.

The logic was to build up the output string in stbuf and once this is bigger
than SC930_DATAVAL_BUFSIZE-10 then write out what we have, reset the buffer and
carry on. Once complete write the remaining string with a new line.

Unfortunately the writing of the 'remainder' part was using a pointer that
had already moved past the end of the string.

Changed adu_sc930prtdataval to properly use stbuf as the pointer to the
remainder of the output string. Also when writing out a segment due to having
exceeded the buffer limit reset to an empty string in case there's no more
of the value to output.

------------------------------------------------------------------------