------------------------------------------------------------------------ r3745 | maspa05 | 2010-10-26 12:45:22 -0400 (Tue, 26 Oct 2010) | 18 lines
Review Board IP  http://reviewboard.ingres.prv/r/6482

Original fix for bug 124551 may not output the whole value for byte, varbyte,
nchar or nvarchar data values.

The logic was to build up the output string in stbuf and once this is bigger
than SC930_DATAVAL_BUFSIZE-10 then write out what we have, reset the buffer and
carry on. Once complete write the remaining string with a new line.

Unfortunately the writing of the 'remainder' part was using a pointer that
had already moved past the end of the string.

Changed adu_sc930prtdataval to properly use stbuf as the pointer to the
remainder of the output string. Also when writing out a segment due to having
exceeded the buffer limit reset to an empty string in case there's no more
of the value to output.

------------------------------------------------------------------------
r3746 | schka01 | 2010-10-26 16:31:42 -0400 (Tue, 26 Oct 2010) | 28 lines

The original motivation was to be able to specify WITH STRUCTURE=HEAP
on a create table, so that in IVW (codev) one could also say WITH
STRUCTURE=VECTORWISE.  The next step would be to tie the SET
RESULT_STRUCTURE statement to allow Vectorwise structures as options,
and replace the existing vectorwise SET / SET SESSION statements.

In doing this, I ran into the old problems related to WITH-parsing:
not only is the parsing scattered around in at least 5 different places,
but also the results are gathered differently depending on statement.
QUEL retrieve into only creates a PST_RESTAB.   MODIFY and CREATE
INDEX and CREATE TABLE only creates a DMU_CB with a DMU
characteristics array.  SQL create-table-as-select creates both.

This change:
a) unifies WITH-clause parsing into a single file (pslwithopts.c) that
is table driven. Option checking is simplified and consistent error
messages are generated.
b) replaces the DMU_CB characteristics array with a DMU_CHARACTERISTICS
structure.  That structure has a bit-map indicating which options
have been supplied, and the values for those options.
The DMU_CHARACTERISTICS is generated by all statements and is
carried through to DMF.
c) Eliminates most of the PST_RESTAB in favor of the DMU
CHARACTERISTICS structure.
d) generates a DMU_CB (and QEU_CB) for all create-like statements,
even QUEL retrieve into, in the parser.

------------------------------------------------------------------------