------------------------------------------------------------------------ r3751 | loera01 | 2010-10-27 16:45:49 -0400 (Wed, 27 Oct 2010) | 3 lines
Added reference to the actual_portID argument for the portmapper function reference in GC_CONN and GC_CONNBK in GCqiosm().  Fixed function comments for GC_

------------------------------------------------------------------------
r3752 | smeke01 | 2010-10-28 06:00:31 -0400 (Thu, 28 Oct 2010) | 29 lines

Review Board IP  http://reviewboard.ingres.prv/r/6495

Cross integration of change 508107 and 508131

When the query contains correlates mapping onto the same var (ie
when ops_gmask has the OPS_IDSAME flag set) we call opa_cvar() on
the candidate aggregate, using a var map from the aggregate being
compared against. However it is possible for the aggregates we are
combining not to match, precisely because they use different vars.
opa_cvar() should only be interested in converting the var numbers
of correlated vars that appear in both the main and candidate
queries. For the rest it should leave the var numbers unchanged,
and not raise an error.

Added Functions:

   None.

Removed Functions:

   None.

CL Interface changes:

   None.

------------------------------------------------------------------------