------------------------------------------------------------------------ r3768 | schka01 | 2010-11-01 16:00:15 -0400 (Mon, 01 Nov 2010) | 10 lines
Description:
Fix a couple additional items related to the WITH-option submission:
- Allow UNIQUE as a with-option in embedded SQL, since it's now allowed
  in the backend as a create table as select WITH-option.
- Change QUEL to take "set result_structure" as well as "set ret_into".
  There was no apparent reason why quel was different.
- Fix a bug in psl_modify_command_string where action enum names were
  incorrect used as DMU indicator names.  Use the right indicator names.

------------------------------------------------------------------------
r3769 | kiria01 | 2010-11-02 08:30:15 -0400 (Tue, 02 Nov 2010) | 28 lines

Scalar Subqueries as DB procedure parameters.

Documentation Impact:

If documented as a restriction - this is lifted.

The intent of this bugfix is to cause any actual parameter
SSQs to be executed before the execute procedure is done.
This requires that temporary variables be assigned to for
passing the results to the procedure. If we are in a DBproc,
this is not a problem as local variables are accessible.
If not in a DBproc we need to extend the DBproc mechanism
to allow a sequence of statements to be executed as one -
one of which will be the DECLARE for temporaries, the other
will be any SSQ query headers and finally the execute itself.

To ease this, the DBP variables list code has been rationalised
into psl_define_var and psl_lookup_var. Similarly, the DBP
epilogue code has been pooled into psl_dbp_finalize and the
prologue code into psl_dbp_init_nondbp and psl_dbp_define_builtins.

This mostly affects pslsgram.yi but OPF modules opcorig.c,
opcqual.c and opsinit.c needed assumptions removing about
DECLARE statements - PST_DV_TYPE can now appear otside of a
DBP context.

------------------------------------------------------------------------