------------------------------------------------------------------------ r3769 | kiria01 | 2010-11-02 08:30:15 -0400 (Tue, 02 Nov 2010) | 28 lines
Scalar Subqueries as DB procedure parameters.

Documentation Impact:

If documented as a restriction - this is lifted.

The intent of this bugfix is to cause any actual parameter
SSQs to be executed before the execute procedure is done.
This requires that temporary variables be assigned to for
passing the results to the procedure. If we are in a DBproc,
this is not a problem as local variables are accessible.
If not in a DBproc we need to extend the DBproc mechanism
to allow a sequence of statements to be executed as one -
one of which will be the DECLARE for temporaries, the other
will be any SSQ query headers and finally the execute itself.

To ease this, the DBP variables list code has been rationalised
into psl_define_var and psl_lookup_var. Similarly, the DBP
epilogue code has been pooled into psl_dbp_finalize and the
prologue code into psl_dbp_init_nondbp and psl_dbp_define_builtins.

This mostly affects pslsgram.yi but OPF modules opcorig.c,
opcqual.c and opsinit.c needed assumptions removing about
DECLARE statements - PST_DV_TYPE can now appear otside of a
DBP context.

------------------------------------------------------------------------
r3770 | pickr01 | 2010-11-02 10:15:22 -0400 (Tue, 02 Nov 2010) | 7 lines

Review Board IP  http://reviewboard.ingres.prv/r/6589

When doing change 508186 two history commands got lost. (for changes 503163 and 507292)
Readded these two comments. Will be integrated in 10.0 too, but not in 9.1, 9.2 or 9.3 as
these changes are not included in those releases.

------------------------------------------------------------------------