------------------------------------------------------------------------ r3780 | hanal04 | 2010-11-05 07:00:15 -0400 (Fri, 05 Nov 2010) | 24 lines
Cross integration of change 508372

Review Board IP  http://reviewboard.ingres.prv/r/6671

When the GCD startup count is altered CBF calls format_port_rollups()
to apply updates to the associated port paramters. The user is
prompted if the port is numeric as the user may not want the update
applied.
iisetres was not making any attempt to updated the associated ports
when the gcd startup count was altered. This change adds
scan_ports_and_change() which is called if we are updaing the
gcd startup count. scan_ports_and_change() is based off of
format_port_rollups() with one minor difference. If the ports
are numeric and the user may not want the ports changed we do not
change the port values and we warn the user that the values
are unchanged. iisetres does not currently prompt users so adding
a prompt may break existing user scripts. If the port values
are changes messages are displayed as follows:

CHANGE ii.hanal0493.gcd.*.tcp_ip.port: (h17)...(h17+)
CHANGE ii.hanal0493.gcd.*.tcp_wol.port: (h17)...(h17+)

------------------------------------------------------------------------