------------------------------------------------------------------------ r3780 | hanal04 | 2010-11-05 07:00:15 -0400 (Fri, 05 Nov 2010) | 24 lines
Cross integration of change 508372

Review Board IP  http://reviewboard.ingres.prv/r/6671

When the GCD startup count is altered CBF calls format_port_rollups()
to apply updates to the associated port paramters. The user is
prompted if the port is numeric as the user may not want the update
applied.
iisetres was not making any attempt to updated the associated ports
when the gcd startup count was altered. This change adds
scan_ports_and_change() which is called if we are updaing the
gcd startup count. scan_ports_and_change() is based off of
format_port_rollups() with one minor difference. If the ports
are numeric and the user may not want the ports changed we do not
change the port values and we warn the user that the values
are unchanged. iisetres does not currently prompt users so adding
a prompt may break existing user scripts. If the port values
are changes messages are displayed as follows:

CHANGE ii.hanal0493.gcd.*.tcp_ip.port: (h17)...(h17+)
CHANGE ii.hanal0493.gcd.*.tcp_wol.port: (h17)...(h17+)

------------------------------------------------------------------------
r3781 | maspa05 | 2010-11-05 07:45:14 -0400 (Fri, 05 Nov 2010) | 28 lines

Review Board IP  http://reviewboard.ingres.prv/r/6675

Add SC930 tracing for XA operations within the DBMS. Update two-phase commit
tracing to make it more useful, including re-formatting of SESSION BEGINS
LINE.

Moved the output of SC930_LTYPE_BEGINTRACE to scs_initiate where we actually
begin the session, rather than in psq_parse where it used to be. It was only
getting output after the first parsed query - which meant it could appear
after XA operations or not at all. Also included the Ingres distributed
transaction id in this line and the database name.

Added sc930_xatrace() to scsqncr.c for SC930 tracing of XA operations. This
is based on scs_xatrace() which is used for an existing trace point. Pulled
out the formatting of the XA distributed transaction id, needed by both and
created scs_format_xa_xid(). Unlike the scs_xatrace we're only interested in
tracing XA operations, not events that occur related to XA. This is analogous
to tracing the fact that an SQL statement has been issued but not all the
consequences of that query. There are other ways to trace this but SC930 is
focussed on query tracing.

For Ingres two-phase commit we already traced the "prepare to commit" statement
but did so using the internal name of "SECURE". Changed this to "PREPCOMMIT"
and added the distributed transaction id.

Increased the SC930_VERSION to 9.

------------------------------------------------------------------------