------------------------------------------------------------------------ r3781 | maspa05 | 2010-11-05 07:45:14 -0400 (Fri, 05 Nov 2010) | 28 lines
Review Board IP  http://reviewboard.ingres.prv/r/6675

Add SC930 tracing for XA operations within the DBMS. Update two-phase commit
tracing to make it more useful, including re-formatting of SESSION BEGINS
LINE.

Moved the output of SC930_LTYPE_BEGINTRACE to scs_initiate where we actually
begin the session, rather than in psq_parse where it used to be. It was only
getting output after the first parsed query - which meant it could appear
after XA operations or not at all. Also included the Ingres distributed
transaction id in this line and the database name.

Added sc930_xatrace() to scsqncr.c for SC930 tracing of XA operations. This
is based on scs_xatrace() which is used for an existing trace point. Pulled
out the formatting of the XA distributed transaction id, needed by both and
created scs_format_xa_xid(). Unlike the scs_xatrace we're only interested in
tracing XA operations, not events that occur related to XA. This is analogous
to tracing the fact that an SQL statement has been issued but not all the
consequences of that query. There are other ways to trace this but SC930 is
focussed on query tracing.

For Ingres two-phase commit we already traced the "prepare to commit" statement
but did so using the internal name of "SECURE". Changed this to "PREPCOMMIT"
and added the distributed transaction id.

Increased the SC930_VERSION to 9.

------------------------------------------------------------------------
r3782 | drivi01 | 2010-11-05 12:30:13 -0400 (Fri, 05 Nov 2010) | 18 lines

PROPAGATE change #508380 from ingres10 to main:

On 64-bit Windows, sep is not working very well.

We're getting a lot of extra logs with "Error,
child did not exit".

This change adds some intelligence to extend wait
time between commands if the initial interval is not
sufficient.  Also, adds additional routines to
kill all the hung processes including the child
processes to avoid abandoned processes from lingering
around.

This reduces the amount of logs left over after handoffqa
tests complete.

------------------------------------------------------------------------