------------------------------------------------------------------------ r3808 | thogo01 | 2010-11-10 14:30:10 -0500 (Wed, 10 Nov 2010) | 19 lines
JDBC driver uses "select substring( <locator> from <position> for <length> )" to
satisfy CharacterStream requests to read the LOB data.  The substring parameters
are character based values, so more data bytes may be returned than requested
when the data is stored in a multi-byte character set like UTF8.

Since the character processing in the driver before being passed to a charset to
Unicode encoder is in bytes, the excess data bytes need to be cached between
requests made by the encoder.

This change adds a simple cache to the LOB byte stream processing class.  The
cache is a vector of data blocks that grows as needed to hold the overflow
data bytes.  When a request is received, data from the cache is first used to
satisfy the request.  If the request is not completly filled from the cache,
then a substring request is made to the server to fill the remainder of the
request.  If more bytes are returned than needed for the request (as may
happen with multi-byte character sets), then the excess are stored in the
cache.

------------------------------------------------------------------------