------------------------------------------------------------------------ r3812 | thogo01 | 2010-11-11 18:15:12 -0500 (Thu, 11 Nov 2010) | 32 lines
JDBC driver currently defaults DATE_ALIAS to 'ingresdate' for backward
compatibility.  This causes two problems.  First, the default conflicts
with the default setting for SEND_INGDATE (which is FALSE).  Second,
VectorWise does not support the ingresdate data type and the default
setting will not work with VW.

Rather than just changing the DATE_ALIAS default to 'ansidate', which
will result in a change of behavior for existing applications, it will
be better to set the default so as to be compatible with the setting
for SEND_INGDATE.  If SEND_INGDATE is true, 'ingresdate' is the best
default for DATE_ALIAS.  If SEND_INGDATE is false, 'ansidate' is the
better default.  For applications using strictly Ingres dates, this
change should not cause a change in behavior.  For applications using
ANSI date/time, this change should provide better behavior.  If this
change results in undesirable behavior for an existing application,
prior behavior can be obtained by explicitly setting the property.

In addition to adapting the DATE_ALIAS default, it makes sense to make
a similar change to SEND_INGDATE.  If SEND_INGDATE is not set and
DATE_ALIAS is, then the default for SEND_INGDATE should be compatible
with the DATE_ALIAS setting.

The resulting behavior will be as follows:

     DATE_ALIAS   SEND_INGDATE
     <unset>      <unset>        DATE_ALIAS ansidate, SEND_INGDATE false
     <unset>      false          DATE_ALIAS ansidate
     <unset>      true           DATE_ALIAS ingresdate
     ansidate     <unset>        SEND_INGDATE false
     ingresdate   <unset>        SEND_INGDATE true

------------------------------------------------------------------------
r3813 | huazh01 | 2010-11-12 14:30:14 -0500 (Fri, 12 Nov 2010) | 27 lines

Review Board IP  http://reviewboard.ingres.prv/r/6735

Cross integration of change 506407

Issue: 145808
bug: 124090

This change is to restrict the fix to b123820, we now
apply the fix to b123820 only if the previous display
attribute is reverse video.

Added Functions:

   None.

Removed Functions:

   None.

CL Interface changes:

   None.

------------------------------------------------------------------------