------------------------------------------------------------------------ r3813 | huazh01 | 2010-11-12 14:30:14 -0500 (Fri, 12 Nov 2010) | 27 lines
Review Board IP  http://reviewboard.ingres.prv/r/6735

Cross integration of change 506407

Issue: 145808
bug: 124090

This change is to restrict the fix to b123820, we now
apply the fix to b123820 only if the previous display
attribute is reverse video.

Added Functions:

   None.

Removed Functions:

   None.

CL Interface changes:

   None.

------------------------------------------------------------------------
r3814 | huazh01 | 2010-11-12 14:30:25 -0500 (Fri, 12 Nov 2010) | 35 lines

Review Board IP  http://reviewboard.ingres.prv/r/6734

Cross integration of change 505306

issue: 143155
bug: 123419

If a query contains an ifnull() on the where clauses part
of a user query, optimizer generates a plan that uses Cart-Prod
and does not join on the column specified in the user query.
This makes the query runs longer.

Rewrite the fix to b117793. On ope_jnclaus(), we now don't
create functional attribute for ifnull() if:

a) ifnull() is in the where clause.
b) variables on both side of the and_node are referenced in
an outer join.

Added Functions:

   None.

Removed Functions:

   None.

CL Interface changes:

   None.

------------------------------------------------------------------------