------------------------------------------------------------------------ r3866 | vande02 | 2010-11-22 18:47:12 -0500 (Mon, 22 Nov 2010) | 9 lines
Cross-integrating MAIN change 507940 to OPENSRCQA.

Error message text for alter table ...was expanded
to include usage message for rename table/columns.
Sep test alttbl21.sep was modified to check for the
new error message. Also sep test alttbl82.sep was
modified to fix the alignment of the error messages.

------------------------------------------------------------------------
r3867 | vande02 | 2010-11-22 18:47:24 -0500 (Mon, 22 Nov 2010) | 14 lines

Cross-integrating MAIN change 508577 to OPENSRCQA.

Test maintenance to LAR.

Due to some Ingres change in release 10 for OpenROAD, ckpdb and
rollforwarddb use one more file during processing on UTF8.  Had to
add that one line and change the File(s) number for several canons.
While I had the files opened, I moved the NT_GENERIC canons to the
top and the general canons to the bottom of the string of canons.
The canons added were specific to this condition:

 << IF (NT_GENERIC && SEPPARAM_CHARSET == "UTF8")

------------------------------------------------------------------------