------------------------------------------------------------------------ r3898 | ingres | 2010-11-29 16:36:43 -0500 (Mon, 29 Nov 2010) | 1 line
Move csinternal.h to complete previous change for bug 124685.
------------------------------------------------------------------------
r3899 | hanal04 | 2010-11-30 05:45:11 -0500 (Tue, 30 Nov 2010) | 16 lines

Cross integration of change 508725

Review Board IP  http://reviewboard.ingres.prv/r/6944

rdd_defsearch() has a control loop in which the RDF_DEF_DESC mutex
may be acquired. When we exit the loop we were check the RDF_RSEMAPHORE
flag to determine whether we needed to release the RDF_DEF_DESC
mutex. This is not the correct flag value. RDF_DSEMAPHORE is set
when we acquire the RDF_DEF_DESC mutex. This change corrects the
IF condition to ensure the correct release of the RDF_DEF_DESC
after we exit the control loop.
This change also backs out the previous changes for Bug 124718 as
they were not a root cause solution.

------------------------------------------------------------------------