------------------------------------------------------------------------ r3904 | vande02 | 2010-11-30 14:30:10 -0500 (Tue, 30 Nov 2010) | 7 lines
Cross-integrating MAIN Change 497721 to OPENSRCQA.

SIR 122145 - Add a qryproc2 suite

This submission creates a new suite, qryproc2, for additional qryproc tests.

------------------------------------------------------------------------
r3905 | vande02 | 2010-11-30 15:30:13 -0500 (Tue, 30 Nov 2010) | 18 lines

Cross-integrating MAIN Change 497835 to OPENSRCQA.

SIR 122145 - Add a qryproc2 suite to

Continuation of SIR 122145.

In the original submission, runbe.* was modified to add/run qryproc2,
either by specifically mentioning the suite, or during 'be all'

A scan of the rest of the test scripts showed the following two
things also need to be updated:

1)  Scripts that run be tests by specifically mentioning each suite.
     In these cases, qryproc2 was added to any script that ran qryproc

2)  Documentation that mentions qryproc needs to also mention qryproc2

------------------------------------------------------------------------