------------------------------------------------------------------------ r3905 | vande02 | 2010-11-30 15:30:13 -0500 (Tue, 30 Nov 2010) | 18 lines
Cross-integrating MAIN Change 497835 to OPENSRCQA.

SIR 122145 - Add a qryproc2 suite to

Continuation of SIR 122145.

In the original submission, runbe.* was modified to add/run qryproc2,
either by specifically mentioning the suite, or during 'be all'

A scan of the rest of the test scripts showed the following two
things also need to be updated:

1)  Scripts that run be tests by specifically mentioning each suite.
     In these cases, qryproc2 was added to any script that ran qryproc

2)  Documentation that mentions qryproc needs to also mention qryproc2

------------------------------------------------------------------------
r3906 | vande02 | 2010-11-30 15:45:12 -0500 (Tue, 30 Nov 2010) | 22 lines

Cross-integrating MAIN Change 497837 to OPENSRCQA.

SIR 122200 - make 'runbe clean all' remove regression test databases.

Some regression tests do not use SEPPARAMDB and instead create
their own database to run.  This will occur if the database requires
a characteristic (often collation sequences) that isn't in the default
database for the suite.

After a run of a sep test, databases and tables are currently left
in place to allow easier debugging of sep test diffs.

This IP modifies the runbe* scripts so that the regression test
databases are removed during a 'runbe clean all'.  The test destroys them by
running a cleanregr.sep test.  Using a sep test makes adding new databases
to the 'destroy list' simpler (rather than continually modifying, and
possibly breaking, the runbe* scripts)

This submission also destroys b58324 which is not a regression test, but
was a lar test that wasn't cleaned up properly.

------------------------------------------------------------------------