------------------------------------------------------------------------ r3906 | vande02 | 2010-11-30 15:45:12 -0500 (Tue, 30 Nov 2010) | 22 lines
Cross-integrating MAIN Change 497837 to OPENSRCQA.

SIR 122200 - make 'runbe clean all' remove regression test databases.

Some regression tests do not use SEPPARAMDB and instead create
their own database to run.  This will occur if the database requires
a characteristic (often collation sequences) that isn't in the default
database for the suite.

After a run of a sep test, databases and tables are currently left
in place to allow easier debugging of sep test diffs.

This IP modifies the runbe* scripts so that the regression test
databases are removed during a 'runbe clean all'.  The test destroys them by
running a cleanregr.sep test.  Using a sep test makes adding new databases
to the 'destroy list' simpler (rather than continually modifying, and
possibly breaking, the runbe* scripts)

This submission also destroys b58324 which is not a regression test, but
was a lar test that wasn't cleaned up properly.

------------------------------------------------------------------------
r3907 | rajus01 | 2010-11-30 16:00:11 -0500 (Tue, 30 Nov 2010) | 3 lines

Crossing the change 508589 from ingres10 - Updated the Ingres JDBC JavaARchive from the ingres10 codeline. It is OK to copy the jar file as the driver source in in the main and ingres10 code lines are identical at this point in time.

------------------------------------------------------------------------