------------------------------------------------------------------------ r3907 | rajus01 | 2010-11-30 16:00:11 -0500 (Tue, 30 Nov 2010) | 3 lines
Crossing the change 508589 from ingres10 - Updated the Ingres JDBC JavaARchive from the ingres10 codeline. It is OK to copy the jar file as the driver source in in the main and ingres10 code lines are identical at this point in time.

------------------------------------------------------------------------
r3908 | vande02 | 2010-11-30 16:15:12 -0500 (Tue, 30 Nov 2010) | 5 lines

Cross-integrating MAIN Change 506111 to OPENSRCQA.

SIR 124088: Add a persistently journaled database for qryproc2.

------------------------------------------------------------------------