------------------------------------------------------------------------ r3912 | vande02 | 2010-11-30 17:30:09 -0500 (Tue, 30 Nov 2010) | 5 lines
Cross-integrating MAIN Change 506850 to OPENSRCQA.

Fix typo in last change.

------------------------------------------------------------------------
r3913 | crogr01 | 2010-11-30 17:44:35 -0500 (Tue, 30 Nov 2010) | 6 lines

This change is the first import of the Ruby Frequent Flyer application into piccolo.
It should be noted that the application does not work with the latest Ruby driver.
However this code is being being checked in so that it is available centrally and
under version control.

------------------------------------------------------------------------