------------------------------------------------------------------------ r3913 | crogr01 | 2010-11-30 17:44:35 -0500 (Tue, 30 Nov 2010) | 6 lines
This change is the first import of the Ruby Frequent Flyer application into piccolo.
It should be noted that the application does not work with the latest Ruby driver.
However this code is being being checked in so that it is available centrally and
under version control.

------------------------------------------------------------------------
r3914 | crogr01 | 2010-11-30 17:46:13 -0500 (Tue, 30 Nov 2010) | 21 lines

Update database_connection.rb:
- Add documentation to each method that accesses the database
- Remove redundant/commented out code
- Improve the handling of connections when a user name is
  provided without a password
Update fp_constants.rb
- Improve the display of text by making the font sizes more
  readable on larger screens
- Set defaultDBConfigPath to "." if the environment variables
  TMP or TEMP are not defined
Update profile_frames.rb
- Handle a condition where by a Nil/Null email address does
  not cause an exception

Add a script and batch file to allow a user to start up the demo

Generated RDoc documentation using the code comments added in
database_connection.rb. A sample copy of the docs can be found at
http://dl.dropbox.com/u/2191213/ruby_ff_demo/doc/index.html

------------------------------------------------------------------------