------------------------------------------------------------------------ r3940 | jenjo02 | 2010-12-03 12:46:37 -0500 (Fri, 03 Dec 2010) | 40 lines
Creates persistent managed objects for a database .cnf file.

As the config representation (DM0C_CNF) comes and goes with
each dm0c_open(), dm0c_close(), it's neccessary to allocate
and maintain a durable set of MO objects while a database
exists in a server (dm2d_add_db()) through dm2d_del_db()).

This set of objects is anchored in the database's DMP_DCB
dcb_mo.

The MO objects are made visible by new IMA tables,
each corresponding to the structure(s):

    o ima_db_config	DM0C_DSC, DM0C_JNL, DM0C_DMP
    o ima_db_extent	DM0C_EXT, DMP_LOC_ENTRY
    o ima_db_jckp	_JNL_CKP
    o ima_db_jnode	_JNL_CNODE_INFO
    o ima_db_dckp	_DMP_CKP
    o ima_db_dnode	_DMP_CNODE_INFO

When the database is deleted from the server, its objects
are detached and no longer visible.

All objects are keyed on the database's database id (dbid)

Though the objects are valued at dm2d_add_db() time, an
IMA procedure is provided to recreate the objects using
the now-current .cnf file:

    execute procedure ima_db_refresh(i4 dbid)

If the <dbid> is not known to IMA (has not been opened
in the server) a status of 0326 is returned, otherwise
one may assume that the objects have been refreshed.

The "refresh" is initiated by a MO set method which in turn
launches a factotum thread to detach the current set of
objects, open the .cnf file, and attach a new set of objects.

------------------------------------------------------------------------
r3941 | schka01 | 2010-12-03 13:30:56 -0500 (Fri, 03 Dec 2010) | 7 lines

Prototype and warning fixes for QEF.
Mostly mechanical;  the most interesting discoveries were
some array overruns in qelsel.c, and the improper use of
CMcpychar to copy single-character ascii constants (which is not
correct, as it references past the end of the constant string.)

------------------------------------------------------------------------